refactor and fixes
* fixes issue when player sprite was made transparent
* for map, changes i32 -> usize, all map positions are cached in vector, so they will always be positive
* removed commented out code

+++ src/systems/map/rect.rs
@@ -1,13 +1,13 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Rect {
-    pub x1: i32,
-    pub x2: i32,
-    pub y1: i32,
-    pub y2: i32,
+    pub x1: usize,
+    pub x2: usize,
+    pub y1: usize,
+    pub y2: usize,
 }
 
 impl Rect {
-    pub fn new(x: i32, y: i32, w: i32, h: i32) -> Rect {
+    pub fn new(x: usize, y: usize, w: usize, h: usize) -> Rect {
         Rect {
             x1: x,
             y1: y,
@@ -21,7 +21,7 @@ impl Rect {
         self.x1 <= other.x2 && self.x2 >= other.x1 && self.y1 <= other.y2 && self.y2 >= other.y1
     }
 
-    pub fn center(&self) -> (i32, i32) {
+    pub fn center(&self) -> (usize, usize) {
         ((self.x1 + self.x2) / 2, (self.y1 + self.y2) / 2)
     }
 }
